api-provider/chain-api-data-agnostic/requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
idna==3.6
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
gunicorn==21.2.0
MarkupSafe==2.1.3
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
Werkzeug==3.0.1
